Splitting DxProxy into 2 projects. DxProxy still contains all the code and build a static library. DxProxyDLL builds a DLL from the static lib. Didn't want to use batch build and msbuild / manual lib.exe tasks were significantly more complicated to set up.
apihijack is built against the static lib and the apihijack.dll is the d3d9.dll replacement when that hooking method is used, dynamic linking was problematic as it was possible for apihijack to end up trying to access methods in the actual d3d9.dll when the only existed in the proxy. (This was breaking Borderlands 2). This was the quickest way to fix things bck to how they worked before. Will be revisited when testing easyhook.